+++ src/renderer/components/helpers/ChatMethods.ts
@@ -252,7 +252,11 @@ export async function sendMessage(chatId: number, message: sendMessageParams) {
   chatStore.effect.jumpToMessage(id, false)
 }
 
-export async function forwardMessage(accountId: number, messageId: number, chatId: number) {
+export async function forwardMessage(
+  accountId: number,
+  messageId: number,
+  chatId: number
+) {
   await BackendRemote.rpc.forwardMessages(accountId, [messageId], chatId)
 }