+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased][unreleased]
 
+### Added
+
+### Changed
+- forward message dialog now has a title and a confirmation dialog before forwarding
+
+### Fixed
+
 <a id="1_34_1"></a>
 
 ## [1.34.1] - 2022-12-22